# Tidy Composer, improve init performance