This is mostly a clean up/refactoring of the existing logic and behavior.
The only real behavior change is not waiting for the CompositionContainer
to be fully resolved before finishing the init procedures. 90% of the calls
we make to the Composer are QC types and therefore handled by the special
case code that explicitly loads QC assemblies and caches their types.

Another behavior change was made to really trim down the number of types
that were being cached from the QC assemblies. We were caching ALL public
types that weren't abstract/interface/enum, but since it's a proxy for MEF,
we should really have the [InheritedExport] filter added. Adding this cut
down the number of QC types were caching by over 100x. Another perf gain
was in how we store the cached items. The existing code had the cache in
a list with over 10k types in it and then would search that list doing
an IsAssignableFrom and MatchesTypeName. That's a lot of items to hit.
This commit changes the storage to be a dictionarty keyed by the contract
type. We use the type that declared the [InheritedExport] as well as the
implementation type to support requested exact types, but I'm not sure
that this was required but at first glance it seemed to be supported by
the existing implementation.

This change cuts down LEAN's startup time (defined as the time it takes
to go from the top of Main to just following the initialization/resolution
of the algorithm handlers. This section of code's performance is 100%
dominated by the Composer's initialization and resolution functions. Avg
times on 'master' were around 350ms for a hot start and over 1.5s for a
cold start. Cold start meaning first run after fresh build. This change
shows an average hot start time of about 210ms and a cold start time of
about 250ms

This change also adds a Composer.HardReset() method which clears all of
the state within the Composer and re-initializes everything, including
reading assemblies from disk.

Closes #3836


<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->

#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->

#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->

#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Refactor (non-breaking change which improves implementation)
- [ ] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] My code follows the code style of this project.
- [ ] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [ ] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [ ] All new and existing tests passed.
- [ ] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->